Fix communal binstubs
Adding a non-default gem home to the gemrc forces RubyGems to place
binstubs there rather than in the main install (even though the default
gem home symlinks to the same place).  Then, all we need to do is force
"#!/usr/bin/env ruby" as the shebang.  Our solution to that only works
if the user doesn't have an "install" line in ~/.gemrc.

Closes #3.